requirements.txt
# FILE: requirements.txt

# Hugging Face ecosystem
# transformers>=4.30.0 # CAUSES ERROR DESPITE HAVING PYTORCH 2.6.0
transformers==4.41.2
tokenizers>=0.13.0
datasets>=2.12.0
accelerate>=0.20.0

# PyTorch Lightning
pytorch-lightning>=2.0.0
lightning>=2.0.0
torchmetrics>=0.11.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Experiment tracking
wandb>=0.15.0
tensorboard

# Evaluation metrics
rouge-score>=0.1.2
rouge-metric
nltk>=3.8.0

# Korean NLP support
konlpy>=0.6.0
mecab-python3>=1.0.6
python-mecab-ko>=1.3.0

# Utilities
python-dotenv>=1.0.0
pathlib2
typing-extensions>=4.5.0

# CLI and debugging utilities
fire>=0.5.0
icecream>=2.1.0
click
rich
pandas
scikit-learn
pyyaml
tqdm
pre-commit

# Data visualization
matplotlib
seaborn==0.13.2